predictive-maintenance-system/edge-agent/requirements.txt
# Edge Agent dependencies
# Python 3.12+

# Web framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Settings
pydantic==2.10.4
pydantic-settings==2.7.0

# ML inference
torch==2.5.1
numpy==2.1.3

# Data processing
pandas==2.2.3
scikit-learn==1.6.0

# Kafka
kafka-python-ng==2.2.3

# OPC-UA (optional, for production)
# asyncua==1.1.5

# Utilities
python-dotenv==1.0.1
